+++ src/Events/BeforeGenerate.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\SiteGenerator\Events;
+
+final readonly class BeforeGenerate
+{
+}