+++ src/Events/PageGenerated.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\SiteGenerator\Events;
+
+use Nexendrie\SiteGenerator\Generator;
+
+final readonly class PageGenerated
+{
+    public function __construct(public string $html, public Generator $generator, public string $filename)
+    {
+    }
+}